refactor: move construct virtual data to loader so utils import faster.
